+++ src/release_info.zig
@@ -1,1 +1,1 @@
-pub const semver = "0.2.5809";
+pub const semver = "0.2.5810";